+++ components/src/core/Drawer/DrawerSubheader.tsx
@@ -18,4 +18,3 @@ DrawerSubheader.displayName = 'DrawerSubheader';
 DrawerSubheader.propTypes = {
     drawerOpen: PropTypes.bool,
 };
-