+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 ## v3.0.0
-- Adds support for nested items in the Drawer component; adjusted Drawer API.
+- Adds support for nested items in the Drawer component; 
+- Additional styling props added to Drawer; 
 
 **Breaking Changes:** 
 - A few props got renamed to avoid further ambiguities: